Reduce boilerplate code in action creators and reducers. Add Transaction action creators and reducer.

+++ assets/src/js/actions/accounts.js
@@ -7,14 +7,17 @@ const getAccountsRequest = makeActionCreator(ActionTypes.GET_ACCOUNTS_REQUEST);
 const getAccountsResponse = makeActionCreator(ActionTypes.GET_ACCOUNTS_RESPONSE, 'json');
 const getAccountsFailed = makeActionCreator(ActionTypes.GET_ACCOUNTS_FAILED, 'errors');
 
-export function getAccounts(token) {
-    return dispatch => {
+export function getAccounts() {
+    return (dispatch, getState) => {
         dispatch(getAccountsRequest());
 
-        const service = new APIService(token);
-        return service.getCollection(APIEndpoints.COLLECTION)
-            .then(response => dispatch(getAccountsResponse(response.data)))
-            .catch(errors => dispatch(getAccountsFailed(errors.data)));
+        const { session } = getState();
+        if (session.accessToken && session.accessToken.isValid()) {
+            const service = new APIService(session.accessToken.value);
+            return service.getCollection(APIEndpoints.COLLECTION)
+                .then(response => dispatch(getAccountsResponse(response.data)))
+                .catch(errors => dispatch(getAccountsFailed(errors.data)));
+        }
     };
 }
 
@@ -22,14 +25,17 @@ const createAccountRequest = makeActionCreator(ActionTypes.CREATE_ACCOUNT_REQUES
 const createAccountResponse = makeActionCreator(ActionTypes.CREATE_ACCOUNT_RESPONSE, 'json');
 const createAccountFailed = makeActionCreator(ActionTypes.CREATE_ACCOUNT_FAILED, 'payload', 'errors');
 
-export function createAccount(token, payload) {
-    return dispatch => {
+export function createAccount(payload) {
+    return (dispatch, getState) => {
         dispatch(createAccountRequest(payload));
 
-        const service = new APIService(token);
-        return service.createResource(APIEndpoints.COLLECTION, payload)
-            .then(response => dispatch(createAccountResponse(response.data)))
-            .catch(errors => dispatch(createAccountFailed(payload, errors.data)));
+        const { session } = getState();
+        if (session.accessToken && session.accessToken.isValid()) {
+            const service = new APIService(session.accessToken.value);
+            return service.createResource(APIEndpoints.COLLECTION, payload)
+                .then(response => dispatch(createAccountResponse(response.data)))
+                .catch(errors => dispatch(createAccountFailed(payload, errors.data)));
+        }
     };
 }
 
@@ -37,14 +43,17 @@ const editAccountRequest = makeActionCreator(ActionTypes.EDIT_ACCOUNT_REQUEST, '
 const editAccountResponse = makeActionCreator(ActionTypes.EDIT_ACCOUNT_RESPONSE, 'account', 'json');
 const editAccountFailed = makeActionCreator(ActionTypes.EDIT_ACCOUNT_FAILED, 'account', 'payload', 'errors');
 
-export function editAccount(token, account, payload) {
-    return dispatch => {
+export function editAccount(account, payload) {
+    return (dispatch, getState) => {
         dispatch(editAccountRequest(account, payload));
 
-        const service = new APIService(token);
-        return service.editResource(APIEndpoints.RESOURCE.replace('{id}', account.id), payload)
-            .then(response => dispatch(editAccountResponse(account, response.data)))
-            .catch(errors => dispatch(editAccountFailed(account, payload, errors.data)));
+        const { session } = getState();
+        if (session.accessToken && session.accessToken.isValid()) {
+            const service = new APIService(session.accessToken.value);
+            return service.editResource(APIEndpoints.RESOURCE.replace('{id}', account.id), payload)
+                .then(response => dispatch(editAccountResponse(account, response.data)))
+                .catch(errors => dispatch(editAccountFailed(account, payload, errors.data)));
+        }
     };
 }
 
@@ -52,13 +61,16 @@ const removeAccountRequest = makeActionCreator(ActionTypes.REMOVE_ACCOUNT_REQUES
 const removeAccountResponse = makeActionCreator(ActionTypes.REMOVE_ACCOUNT_RESPONSE, 'account');
 const removeAccountFailed = makeActionCreator(ActionTypes.REMOVE_ACCOUNT_FAILED, 'account', 'errors');
 
-export function removeAccount(token, account) {
-    return dispatch => {
+export function removeAccount(account) {
+    return (dispatch, getState) => {
         dispatch(removeAccountRequest(account));
 
-        const service = new APIService(token);
-        return service.removeResource(APIEndpoints.RESOURCE.replace('{id}', account.id))
-            .then(() => dispatch(removeAccountResponse(account)))
-            .catch(errors => dispatch(removeAccountFailed(account, errors.data)));
+        const { session } = getState();
+        if (session.accessToken && session.accessToken.isValid()) {
+            const service = new APIService(session.accessToken.value);
+            return service.removeResource(APIEndpoints.RESOURCE.replace('{id}', account.id))
+                .then(() => dispatch(removeAccountResponse(account)))
+                .catch(errors => dispatch(removeAccountFailed(account, errors.data)));
+        }
     };
 }
