+++ assets/src/js/actions/categories.js
@@ -7,14 +7,17 @@ const getCategoriesRequest = makeActionCreator(ActionTypes.GET_CATEGORIES_REQUES
 const getCategoriesResponse = makeActionCreator(ActionTypes.GET_CATEGORIES_RESPONSE, 'json');
 const getCategoriesFailed = makeActionCreator(ActionTypes.GET_CATEGORIES_FAILED, 'errors');
 
-export function getCategories(token) {
-    return dispatch => {
+export function getCategories() {
+    return (dispatch, getState) => {
         dispatch(getCategoriesRequest());
 
-        const service = new APIService(token);
-        return service.getCollection(APIEndpoints.COLLECTION)
-            .then(response => dispatch(getCategoriesResponse(response.data)))
-            .catch(errors => dispatch(getCategoriesFailed(errors.data)));
+        const { session } = getState();
+        if (session.accessToken && session.accessToken.isValid()) {
+            const service = new APIService(session.accessToken.value);
+            return service.getCollection(APIEndpoints.COLLECTION)
+                .then(response => dispatch(getCategoriesResponse(response.data)))
+                .catch(errors => dispatch(getCategoriesFailed(errors.data)));
+        }
     };
 }
 
@@ -22,14 +25,17 @@ const createCategoryRequest = makeActionCreator(ActionTypes.CREATE_CATEGORY_REQU
 const createCategoryResponse = makeActionCreator(ActionTypes.CREATE_CATEGORY_RESPONSE, 'json');
 const createCategoryFailed = makeActionCreator(ActionTypes.CREATE_CATEGORY_FAILED, 'payload', 'errors');
 
-export function createCategory(token, payload) {
-    return dispatch => {
+export function createCategory(payload) {
+    return (dispatch, getState) => {
         dispatch(createCategoryRequest(payload));
 
-        const service = new APIService(token);
-        return service.createResource(APIEndpoints.COLLECTION, payload)
-            .then(response => dispatch(createCategoryResponse(response.data)))
-            .catch(errors => dispatch(createCategoryFailed(payload, errors.data)));
+        const { session } = getState();
+        if (session.accessToken && session.accessToken.isValid()) {
+            const service = new APIService(session.accessToken.value);
+            return service.createResource(APIEndpoints.COLLECTION, payload)
+                .then(response => dispatch(createCategoryResponse(response.data)))
+                .catch(errors => dispatch(createCategoryFailed(payload, errors.data)));
+        }
     };
 }
 
@@ -38,14 +44,17 @@ const editCategoryRequest = makeActionCreator(ActionTypes.EDIT_CATEGORY_REQUEST,
 const editCategoryResponse = makeActionCreator(ActionTypes.EDIT_CATEGORY_RESPONSE, 'category', 'json');
 const editCategoryFailed = makeActionCreator(ActionTypes.EDIT_CATEGORY_FAILED, 'category', 'payload', 'errors');
 
-export function editCategory(token, category, payload) {
-    return dispatch => {
+export function editCategory(category, payload) {
+    return (dispatch, getState) => {
         dispatch(editCategoryRequest(category, payload));
 
-        const service = new APIService(token);
-        return service.editResource(APIEndpoints.RESOURCE.replace('{id}', category.id), payload)
-            .then(response => dispatch(editCategoryResponse(category, response.data)))
-            .catch(errors => dispatch(editCategoryFailed(category, payload, errors.data)));
+        const { session } = getState();
+        if (session.accessToken && session.accessToken.isValid()) {
+            const service = new APIService(session.accessToken.value);
+            return service.editResource(APIEndpoints.RESOURCE.replace('{id}', category.id), payload)
+                .then(response => dispatch(editCategoryResponse(category, response.data)))
+                .catch(errors => dispatch(editCategoryFailed(category, payload, errors.data)));
+        }
     };
 }
 
@@ -54,13 +63,16 @@ const removeCategoryRequest = makeActionCreator(ActionTypes.REMOVE_CATEGORY_REQU
 const removeCategoryResponse = makeActionCreator(ActionTypes.REMOVE_CATEGORY_RESPONSE, 'category');
 const removeCategoryFailed = makeActionCreator(ActionTypes.REMOVE_CATEGORY_FAILED, 'category', 'errors');
 
-export function removeCategory(token, category) {
-    return dispatch => {
+export function removeCategory(category) {
+    return (dispatch, getState) => {
         dispatch(removeCategoryRequest(category));
 
-        const service = new APIService(token);
-        return service.removeResource(APIEndpoints.RESOURCE.replace('{id}', category.id))
-            .then(() => dispatch(removeCategoryResponse(category)))
-            .catch(errors => dispatch(removeCategoryFailed(category, errors.data)));
+        const { session } = getState();
+        if (session.accessToken && session.accessToken.isValid()) {
+            const service = new APIService(session.accessToken.value);
+            return service.removeResource(APIEndpoints.RESOURCE.replace('{id}', category.id))
+                .then(() => dispatch(removeCategoryResponse(category)))
+                .catch(errors => dispatch(removeCategoryFailed(category, errors.data)));
+        }
     };
 }
